подключена настройка смещения камеры.

+++ Registrator/IRB Frame/ShotDesc.cs
@@ -49,10 +49,10 @@ namespace Registrator
         map_object_info _map_object_info;
         map_point_info _map_point_info;
 
-        double m_dist = 0.0;
+        ulong m_dist = 0;
         double _msec = 0;
-        Int16 m_line = 0;
-        Int16 m_path = 0;
+        Int32 m_line = 0;
+        Int32 m_path = 0;
         String m_obj = "";
         String m_pNo = "";
         String m_peregon = "";
@@ -60,11 +60,11 @@ namespace Registrator
         public ShotType TypeOfShot { get { return m_shotType; } set { m_shotType = value; } }
         public map_object_info map_object_info { get { return _map_object_info; } set { _map_object_info = value; } }
         public map_point_info map_point_info { get { return _map_point_info; } set { _map_point_info = value; } }
-        public double Distance { get { return m_dist; } set { m_dist = value; } }
+        public ulong Distance { get { return m_dist; } set { m_dist = value; } }
         public String MsecString { get { return irb_frame_time_helper.build_time_string_from_time(_msec); } }
         public double Msec { get { return _msec; } set { _msec = value; } }
-        public Int16 Line { get { return m_line; } set { m_line = value; } }
-        public Int16 Path { get { return m_path; } set { m_path = value; } }
+        public Int32 Line { get { return m_line; } set { m_line = value; } }
+        public Int32 Path { get { return m_path; } set { m_path = value; } }
         public String ObjName { get { return m_obj; } set { m_obj = value; } }
         public String PicketNOffset { get { return m_pNo; } set { m_pNo = value; } }
         public String Peregon { get { return m_peregon; } set { m_peregon = value; } }
